dirstudio/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "fastapi>=0.128.0",
    "imagehash>=4.3.2",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-magic>=0.4.27",
    "python-magic-bin==0.4.14",
    "sqlalchemy>=2.0.45",
]
